Phase 24 — integer-only money everywhere + PaymentStatus enum fix
Every money column flips from decimal(10,2) to UNSIGNED INT. Inputs
step by 1 (no 1.21 etc). Currency formatters drop the fraction.

Schema (24.1). Two migrations alter columns:
  central — subscription_plans.{price_monthly,price_yearly}
  tenant  — payments.{amount,cash_amount,card_amount,insurance_amount},
            clinic_services.price, consultation_service.price_at_time.
Existing demo decimals get floored. clinic_settings.pricing JSON
already stores numbers — no schema work, just read/write casts.

Models (24.2). 'decimal:2' → 'integer' on SubscriptionPlan, Payment,
Service, ConsultationService. BillingCalculator returns ints across
the board (base_price, services[], services_total, total). Test
assertions in CompleteConsultationWithServicesTest + ServiceCrudTest
switch from `(float) ...->toBe(60.0)` to `(int) ...->toBe(60)`.

Form requests (24.3). 'numeric' → 'integer' on every money rule,
range caps drop the .99 fractional portion. StorePaymentRequest
amount min flips from 0.01 → 1. Pricing settings, services,
subscription plans all aligned.

Frontend (24.4). Every step="0.01" on a money input → step="1"
(PaymentForm cash/card/insurance/total, Settings pricing tab,
Services form, Central Plans). Every Intl.NumberFormat with
`currency: 'USD'` adds `maximumFractionDigits: 0` so prices render
as $60 instead of $60.00 across Tenant Dashboard, Pricing landing,
Patients/Show, Payments/Receipt, Doctor/Consultation, Payments/Index,
Reception/Dashboard, Services, Subscription, Reports.

Bonus fix folded in: ReceptionDashboardController used a
PaymentStatus::Succeeded case that doesn't exist on the enum
(catalogue is Pending | Paid | Partial | Refunded). The Pending-
payment filter now whereIn(Paid, Partial); Pending and Refunded
keep the visit waiting for collection. Caught at runtime when
loading /reception.

Verification: tsc clean, pnpm build green (main 359 kB / 117 kB gz),
Pest 110/110, smoke 200 across all authed routes + /reception +
/services + /payments, log empty.

+++ app/Http/Requests/Central/UpdatePlanRequest.php
@@ -20,8 +20,8 @@ class UpdatePlanRequest extends FormRequest
     {
         return [
             'name' => ['required', 'string', 'max:80'],
-            'price_monthly' => ['required', 'numeric', 'min:0'],
-            'price_yearly' => ['required', 'numeric', 'min:0'],
+            'price_monthly' => ['required', 'integer', 'min:0'],
+            'price_yearly' => ['required', 'integer', 'min:0'],
             'max_patients' => ['required', 'integer', 'min:0'],
             'max_staff' => ['required', 'integer', 'min:0'],
             'features' => ['nullable', 'array'],
